Provide option not to write modTime
When using as a go:generate generator, the output should be idempotent.
The modTime on a source file is not relevant in some (all?) situations,
not least because it would differ based on the arbitrary modTime of the
source file itself (which might change between systems). Hence it
carries no information.